feat(audits): add Hermes and React Native codegen audits
- Implement HermesDisabledAudit to check for Hermes being disabled in Podfile, build.gradle, and gradle.properties.
- Add ReactNativeCodegenMissingAudit to flag missing codegenConfig in projects using Turbo Native Modules or Fabric components.
- Introduce navigation and styling audits to detect deprecated APIs and inline styles in React Native projects.
- Create a CLI for managing baseline findings, comparing scan reports, and generating remediation plans.
- Enhance command-line arguments for output formats and severity levels.
