+++ src/audits/code_quality/long_function/brace.rs
@@ -1,90 +1,7 @@
-use crate::audits::traits::FileAudit;
-use crate::findings::types::{Evidence, Finding, FindingCategory, Severity};
-use crate::scan::config::ScanConfig;
-use crate::scan::facts::FileFacts;
+use crate::findings::types::Finding;
 use std::path::Path;
 
-pub struct LongFunctionAudit;
-
-impl FileAudit for LongFunctionAudit {
-    fn audit(&self, file: &FileFacts, config: &ScanConfig) -> Vec<Finding> {
-        if file.content.is_empty() {
-            return vec![];
-        }
-        let language = match file.language.as_deref() {
-            Some(l) if is_supported(l) => l,
-            _ => return vec![],
-        };
-        detect_long_functions(
-            &file.content,
-            language,
-            &file.path,
-            config.long_function_loc_threshold,
-        )
-    }
-}
-
-fn is_supported(language: &str) -> bool {
-    matches!(
-        language,
-        "Rust"
-            | "Go"
-            | "Python"
-            | "TypeScript"
-            | "TypeScript React"
-            | "JavaScript"
-            | "JavaScript React"
-            | "Java"
-            | "Kotlin"
-    )
-}
-
-fn detect_long_functions(
-    content: &str,
-    language: &str,
-    path: &Path,
-    threshold: usize,
-) -> Vec<Finding> {
-    if language == "Python" {
-        detect_python(content, path, threshold)
-    } else {
-        detect_brace_based(content, language, path, threshold)
-    }
-}
-
-fn kotlin_strip_modifiers(trimmed: &str) -> &str {
-    let modifiers = [
-        "private ",
-        "public ",
-        "protected ",
-        "internal ",
-        "open ",
-        "override ",
-        "abstract ",
-        "suspend ",
-        "inline ",
-        "operator ",
-        "external ",
-        "tailrec ",
-        "actual ",
-        "expect ",
-    ];
-    let mut s = trimmed;
-    loop {
-        let before = s;
-        for m in modifiers {
-            s = s.strip_prefix(m).unwrap_or(s);
-        }
-        if s == before {
-            break;
-        }
-    }
-    s
-}
-
-// ── Brace-based detection (Rust, Go, TypeScript, JavaScript) ──────────────────
-
-fn detect_brace_based(
+pub(super) fn detect_brace_based(
     content: &str,
     language: &str,
     path: &Path,
@@ -150,7 +67,7 @@ fn detect_brace_based(
         };
 
         if fn_len > effective_threshold {
-            findings.push(build_finding(
+            findings.push(super::build_finding(
                 path,
                 fn_start_line,
                 fn_end + 1,
@@ -164,6 +81,36 @@ fn detect_brace_based(
     }
 
     findings
+}
+
+fn kotlin_strip_modifiers(trimmed: &str) -> &str {
+    let modifiers = [
+        "private ",
+        "public ",
+        "protected ",
+        "internal ",
+        "open ",
+        "override ",
+        "abstract ",
+        "suspend ",
+        "inline ",
+        "operator ",
+        "external ",
+        "tailrec ",
+        "actual ",
+        "expect ",
+    ];
+    let mut s = trimmed;
+    loop {
+        let before = s;
+        for m in modifiers {
+            s = s.strip_prefix(m).unwrap_or(s);
+        }
+        if s == before {
+            break;
+        }
+    }
+    s
 }
 
 fn is_function_start(trimmed: &str, language: &str) -> bool {
@@ -277,106 +224,4 @@ fn extract_name(trimmed: &str, language: &str) -> String {
 /// positives from `//` inside string literals as an acceptable heuristic cost.
 fn strip_line_comment(line: &str) -> &str {
     line.find("//").map(|pos| &line[..pos]).unwrap_or(line)
-}
-
-// ── Python indentation-based detection ────────────────────────────────────────
-
-fn detect_python(content: &str, path: &Path, threshold: usize) -> Vec<Finding> {
-    let mut findings = Vec::new();
-    let lines: Vec<&str> = content.lines().collect();
-    let total = lines.len();
-
-    // Stack of (start_line_1based, indent_len, fn_name)
-    let mut stack: Vec<(usize, usize, String)> = Vec::new();
-
-    for (i, line) in lines.iter().enumerate() {
-        let stripped = line.trim_start();
-        if stripped.is_empty() || stripped.starts_with('#') {
-            continue;
-        }
-        let indent = line.len() - stripped.len();
-        let line_num = i + 1;
-
-        // Close any functions that have dedented to or past our level
-        while let Some(&(_, fn_indent, _)) = stack.last() {
-            if indent <= fn_indent {
-                let Some((start, _, ref name)) = stack.pop() else {
-                    break;
-                };
-                let fn_len = line_num - start;
-                if fn_len > threshold {
-                    findings.push(build_finding(
-                        path,
-                        start,
-                        line_num - 1,
-                        name,
-                        fn_len,
-                        threshold,
-                    ));
-                }
-            } else {
-                break;
-            }
-        }
-
-        if stripped.starts_with("def ") && stripped.contains('(') {
-            let rest = stripped.trim_start_matches("def ").trim();
-            let end = rest.find('(').unwrap_or(rest.len());
-            let fn_name = rest[..end].trim().to_string();
-            stack.push((line_num, indent, fn_name));
-        }
-    }
-
-    // Flush functions that extend to end of file
-    for (start, _, name) in stack {
-        let fn_len = total + 1 - start;
-        if fn_len > threshold {
-            findings.push(build_finding(path, start, total, &name, fn_len, threshold));
-        }
-    }
-
-    findings
-}
-
-// ── Finding construction ──────────────────────────────────────────────────────
-
-fn build_finding(
-    path: &Path,
-    start_line: usize,
-    end_line: usize,
-    fn_name: &str,
-    fn_len: usize,
-    threshold: usize,
-) -> Finding {
-    let (title, name_display) = if fn_name.is_empty() {
-        (
-            "Long inline function".to_string(),
-            "inline function".to_string(),
-        )
-    } else {
-        (
-            format!("Long function: `{fn_name}`"),
-            format!("`{fn_name}`"),
-        )
-    };
-
-    Finding {
-        id: String::new(),
-        rule_id: "code-quality.long-function".to_string(),
-        title,
-        description: format!(
-            "Function {name_display} spans {fn_len} lines, exceeding the {threshold}-line threshold. \
-             Long functions are harder to test and reason about — consider extracting helper functions."
-        ),
-        category: FindingCategory::CodeQuality,
-        severity: Severity::Medium,
-        evidence: vec![Evidence {
-            path: path.to_path_buf(),
-            line_start: start_line,
-            line_end: Some(end_line),
-            snippet: format!("function spans lines {start_line}–{end_line} ({fn_len} lines)"),
-        }],
-        workspace_package: None,
-        docs_url: None,
-    }
 }
