+++ src/cli/mod.rs
@@ -1,7 +1,10 @@
-use clap::{Parser, Subcommand, ValueEnum};
-use repopilot::baseline::gate::FailOn;
-use repopilot::findings::types::Severity;
-use repopilot::output::OutputFormat;
+mod args;
+
+pub use args::{
+    parse_vibe_budget, CompareOutputFormatArg, FailOnArg, OutputFormatArg, SeverityArg,
+};
+
+use clap::{Parser, Subcommand};
 use std::path::PathBuf;
 
 #[derive(Parser)]
@@ -67,17 +70,10 @@ repopilot compare before.json after.json --format markdown\n  \
 repopilot compare before.json after.json --format json --output diff.json"
     )]
     Compare {
-        /// Path to the earlier scan report (JSON)
         before: std::path::PathBuf,
-
-        /// Path to the more recent scan report (JSON)
         after: std::path::PathBuf,
-
-        /// Output format
         #[arg(long, value_enum, default_value = "console")]
         format: CompareOutputFormatArg,
-
-        /// Write report to a file instead of stdout
         #[arg(short, long)]
         output: Option<std::path::PathBuf>,
     },
@@ -107,54 +103,29 @@ repopilot scan . --baseline .repopilot/baseline.json --fail-on new-high\n  \
 repopilot scan . --max-file-loc 500 --max-directory-modules 30"
     )]
     Scan {
-        /// Path to project, folder, or file
         path: PathBuf,
-
-        /// Output format (console, json, markdown, html, sarif)
         #[arg(long, value_enum)]
         format: Option<OutputFormatArg>,
-
-        /// Write report to a file instead of stdout
         #[arg(short, long)]
         output: Option<PathBuf>,
-
-        /// Path to a RepoPilot config file
         #[arg(long)]
         config: Option<PathBuf>,
-
-        /// Path to a RepoPilot baseline file
         #[arg(long)]
         baseline: Option<PathBuf>,
-
-        /// Fail with exit code 1 when findings meet the selected threshold
         #[arg(long, value_enum)]
         fail_on: Option<FailOnArg>,
-
-        /// Maximum non-empty LOC before a file is reported as large (default: 300)
         #[arg(long)]
         max_file_loc: Option<usize>,
-
-        /// Maximum number of files in a single directory before flagging (default: 20)
         #[arg(long)]
         max_directory_modules: Option<usize>,
-
-        /// Maximum directory nesting depth before flagging (default: 5)
         #[arg(long)]
         max_directory_depth: Option<usize>,
-
-        /// Scan each workspace package separately and group findings by package
         #[arg(long, short = 'w')]
         workspace: bool,
-
-        /// Only show findings at or above this severity level
         #[arg(long, value_enum)]
         min_severity: Option<SeverityArg>,
-
-        /// Print scan phase timing breakdown after the report
         #[arg(long)]
         verbose: bool,
-
-        /// Apply a threshold preset: strict, balanced (default), lenient
         #[arg(long, value_parser = ["strict", "balanced", "lenient"])]
         preset: Option<String>,
     },
@@ -181,51 +152,28 @@ repopilot review . --baseline .repopilot/baseline.json --fail-on new-high\n  \
 repopilot review . --format json --output review.json"
     )]
     Review {
-        /// Path to project, folder, or file
         #[arg(default_value = ".")]
         path: PathBuf,
-
-        /// Base Git ref for review diff. Without this, review compares the working tree against HEAD
         #[arg(long)]
         base: Option<String>,
-
-        /// Head Git ref for review diff. Requires --base and defaults to HEAD when --base is set
         #[arg(long)]
         head: Option<String>,
-
-        /// Path to a RepoPilot config file
         #[arg(long)]
         config: Option<PathBuf>,
-
-        /// Path to a RepoPilot baseline file
         #[arg(long)]
         baseline: Option<PathBuf>,
-
-        /// Fail with exit code 1 when in-diff findings meet the selected threshold
         #[arg(long, value_enum)]
         fail_on: Option<FailOnArg>,
-
-        /// Output format (console, json, markdown)
         #[arg(long, value_enum, default_value = "console")]
         format: CompareOutputFormatArg,
-
-        /// Write report to a file instead of stdout
         #[arg(short, long)]
         output: Option<PathBuf>,
-
-        /// Maximum non-empty LOC before a file is reported as large (default: 300)
         #[arg(long)]
         max_file_loc: Option<usize>,
-
-        /// Maximum number of files in a single directory before flagging (default: 20)
         #[arg(long)]
         max_directory_modules: Option<usize>,
-
-        /// Maximum directory nesting depth before flagging (default: 5)
         #[arg(long)]
         max_directory_depth: Option<usize>,
-
-        /// Only show findings at or above this severity level
         #[arg(long, value_enum)]
         min_severity: Option<SeverityArg>,
     },
@@ -249,26 +197,15 @@ repopilot vibe . --output vibe.md\n  \
 repopilot vibe . --no-header | pbcopy"
     )]
     Vibe {
-        /// Path to project, folder, or file
         path: PathBuf,
-
-        /// Path to a RepoPilot config file
         #[arg(long)]
         config: Option<PathBuf>,
-
-        /// Limit output to a single category: security, arch, architecture, quality, framework, all
         #[arg(long, value_parser = ["security", "arch", "architecture", "quality", "framework", "all"])]
         focus: Option<String>,
-
-        /// Target token budget for output: 2k, 4k, 8k, 16k (default: 4k)
         #[arg(long, value_parser = parse_vibe_budget)]
         budget: Option<usize>,
-
-        /// Write output to a file instead of stdout
         #[arg(short, long)]
         output: Option<PathBuf>,
-
-        /// Omit the intro header block (useful for piping into an LLM API)
         #[arg(long)]
         no_header: bool,
     },
@@ -288,22 +225,13 @@ repopilot harden . --budget 8k\n  \
 repopilot harden . --output harden.md"
     )]
     Harden {
-        /// Path to project, folder, or file
         path: PathBuf,
-
-        /// Path to a RepoPilot config file
         #[arg(long)]
         config: Option<PathBuf>,
-
-        /// Limit output to a single category: security, arch, architecture, quality, framework, all
         #[arg(long, value_parser = ["security", "arch", "architecture", "quality", "framework", "all"])]
         focus: Option<String>,
-
-        /// Target token budget for output: 2k, 4k, 8k, 16k (default: 4k)
         #[arg(long, value_parser = parse_vibe_budget)]
         budget: Option<usize>,
-
-        /// Write output to a file instead of stdout
         #[arg(short, long)]
         output: Option<PathBuf>,
     },
@@ -322,22 +250,13 @@ repopilot prompt . --focus security --budget 2k\n  \
 repopilot prompt . --output prompt.md"
     )]
     Prompt {
-        /// Path to project, folder, or file
         path: PathBuf,
-
-        /// Path to a RepoPilot config file
         #[arg(long)]
         config: Option<PathBuf>,
-
-        /// Limit output to a single category: security, arch, architecture, quality, framework, all
         #[arg(long, value_parser = ["security", "arch", "architecture", "quality", "framework", "all"])]
         focus: Option<String>,
-
-        /// Target token budget for embedded context: 2k, 4k, 8k, 16k (default: 4k)
         #[arg(long, value_parser = parse_vibe_budget)]
         budget: Option<usize>,
-
-        /// Write output to a file instead of stdout
         #[arg(short, long)]
         output: Option<PathBuf>,
     },
@@ -355,11 +274,8 @@ repopilot init --force            # overwrite existing config\n  \
 repopilot init --path ./config/repopilot.toml"
     )]
     Init {
-        /// Overwrite an existing config file
         #[arg(long)]
         force: bool,
-
-        /// Path where the config file should be written
         #[arg(long, default_value = "repopilot.toml")]
         path: PathBuf,
     },
@@ -384,107 +300,10 @@ repopilot baseline create . --output ./baseline.json\n  \
 repopilot baseline create . --force"
     )]
     Create {
-        /// Path to project, folder, or file
         path: PathBuf,
-
-        /// Write baseline to a custom path (default: .repopilot/baseline.json)
         #[arg(short, long)]
         output: Option<PathBuf>,
-
-        /// Overwrite an existing baseline file
         #[arg(long)]
         force: bool,
     },
-}
-
-#[derive(Clone, Copy, Debug, ValueEnum)]
-pub enum OutputFormatArg {
-    Console,
-    Html,
-    Json,
-    Markdown,
-    Sarif,
-}
-
-#[derive(Clone, Copy, Debug, ValueEnum)]
-pub enum CompareOutputFormatArg {
-    Console,
-    Json,
-    Markdown,
-}
-
-#[derive(Clone, Copy, Debug, ValueEnum)]
-pub enum SeverityArg {
-    Info,
-    Low,
-    Medium,
-    High,
-    Critical,
-}
-
-#[derive(Clone, Copy, Debug, ValueEnum)]
-pub enum FailOnArg {
-    NewLow,
-    NewMedium,
-    NewHigh,
-    NewCritical,
-    Low,
-    Medium,
-    High,
-    Critical,
-}
-
-impl From<CompareOutputFormatArg> for OutputFormat {
-    fn from(format: CompareOutputFormatArg) -> Self {
-        match format {
-            CompareOutputFormatArg::Console => OutputFormat::Console,
-            CompareOutputFormatArg::Json => OutputFormat::Json,
-            CompareOutputFormatArg::Markdown => OutputFormat::Markdown,
-        }
-    }
-}
-
-impl From<OutputFormatArg> for OutputFormat {
-    fn from(format: OutputFormatArg) -> Self {
-        match format {
-            OutputFormatArg::Console => OutputFormat::Console,
-            OutputFormatArg::Html => OutputFormat::Html,
-            OutputFormatArg::Json => OutputFormat::Json,
-            OutputFormatArg::Markdown => OutputFormat::Markdown,
-            OutputFormatArg::Sarif => OutputFormat::Sarif,
-        }
-    }
-}
-
-impl From<FailOnArg> for FailOn {
-    fn from(value: FailOnArg) -> Self {
-        match value {
-            FailOnArg::NewLow => FailOn::New(Severity::Low),
-            FailOnArg::NewMedium => FailOn::New(Severity::Medium),
-            FailOnArg::NewHigh => FailOn::New(Severity::High),
-            FailOnArg::NewCritical => FailOn::New(Severity::Critical),
-            FailOnArg::Low => FailOn::Any(Severity::Low),
-            FailOnArg::Medium => FailOn::Any(Severity::Medium),
-            FailOnArg::High => FailOn::Any(Severity::High),
-            FailOnArg::Critical => FailOn::Any(Severity::Critical),
-        }
-    }
-}
-
-fn parse_vibe_budget(value: &str) -> Result<usize, String> {
-    let tokens = match value {
-        "2k" => 2048,
-        "4k" => 4096,
-        "8k" => 8192,
-        "16k" => 16384,
-        other => other
-            .parse::<usize>()
-            .map_err(|_| "expected 2k, 4k, 8k, 16k, or a positive token count".to_string())?,
-    };
-
-    if tokens == 0 {
-        return Err("budget must be greater than zero".to_string());
-    }
-
-    Ok(tokens)
 }
