+++ tests/compare_diff.rs
@@ -76,6 +76,7 @@ fn summary(findings: Vec<Finding>) -> ScanSummary {
         react_native: None,
         coupling_graph: None,
         scan_duration_us: 0,
+        health_score: 0,
     }
 }
