+++ tests/output_html.rs
@@ -34,6 +34,7 @@ fn html_output_escapes_snippets_and_renders_summary() {
         react_native: None,
         coupling_graph: None,
         scan_duration_us: 0,
+        health_score: 0,
     };
 
     let html = render_scan_summary(&summary, OutputFormat::Html).expect("failed to render html");
