+++ tests/output_json.rs
@@ -22,6 +22,7 @@ fn renders_valid_json_scan_summary() {
         react_native: None,
         coupling_graph: None,
         scan_duration_us: 0,
+        health_score: 0,
     };
 
     let output =
