+++ tests/output_markdown.rs
@@ -45,6 +45,7 @@ fn renders_markdown_scan_summary() {
         react_native: None,
         coupling_graph: None,
         scan_duration_us: 0,
+        health_score: 0,
     };
 
     let output = render_scan_summary(&summary, OutputFormat::Markdown)
@@ -79,6 +80,7 @@ fn renders_empty_markdown_sections() {
         react_native: None,
         coupling_graph: None,
         scan_duration_us: 0,
+        health_score: 0,
     };
 
     let output = render_scan_summary(&summary, OutputFormat::Markdown)
@@ -121,6 +123,7 @@ fn renders_react_native_architecture_section_when_profile_present() {
         }),
         coupling_graph: None,
         scan_duration_us: 0,
+        health_score: 0,
     };
 
     let output = render_scan_summary(&summary, OutputFormat::Markdown)
