+++ tests/scan_summary.rs
@@ -1,6 +1,8 @@
+use repopilot::findings::types::{Finding, FindingCategory, Severity};
 use repopilot::scan::config::ScanConfig;
 use repopilot::scan::scanner::scan_path;
 use repopilot::scan::scanner::scan_path_with_config;
+use repopilot::scan::types::ScanSummary;
 use std::fs;
 use tempfile::tempdir;
 
@@ -89,4 +91,70 @@ fn scan_reports_binary_files_as_skipped_without_failing() {
     assert_eq!(summary.skipped_files_count, 1);
     assert_eq!(summary.skipped_bytes, bytes.len() as u64);
     assert_eq!(summary.lines_of_code, 0);
+}
+
+#[test]
+fn health_score_is_100_with_no_findings() {
+    assert_eq!(ScanSummary::compute_health_score(&[], 1000), 100);
+}
+
+#[test]
+fn health_score_decreases_with_severity() {
+    fn finding(severity: Severity) -> Finding {
+        Finding {
+            id: String::new(),
+            rule_id: "test".to_string(),
+            title: String::new(),
+            description: String::new(),
+            category: FindingCategory::Security,
+            severity,
+            evidence: vec![],
+            workspace_package: None,
+            docs_url: None,
+        }
+    }
+    let critical = ScanSummary::compute_health_score(&[finding(Severity::Critical)], 1000);
+    let high = ScanSummary::compute_health_score(&[finding(Severity::High)], 1000);
+    let medium = ScanSummary::compute_health_score(&[finding(Severity::Medium)], 1000);
+    let low = ScanSummary::compute_health_score(&[finding(Severity::Low)], 1000);
+
+    assert!(critical < high, "critical should score worse than high");
+    assert!(high < medium, "high should score worse than medium");
+    assert!(medium < low, "medium should score worse than low");
+    assert!(low < 100, "any finding should reduce score from 100");
+}
+
+#[test]
+fn health_score_is_clamped_at_zero_for_catastrophic_repos() {
+    let findings: Vec<Finding> = (0..10)
+        .map(|_| Finding {
+            id: String::new(),
+            rule_id: "test".to_string(),
+            title: String::new(),
+            description: String::new(),
+            category: FindingCategory::Security,
+            severity: Severity::Critical,
+            evidence: vec![],
+            workspace_package: None,
+            docs_url: None,
+        })
+        .collect();
+    assert_eq!(
+        ScanSummary::compute_health_score(&findings, 100),
+        0,
+        "score must clamp to 0"
+    );
+}
+
+#[test]
+fn scan_result_includes_health_score() {
+    let temp = tempdir().expect("failed to create temp dir");
+    fs::write(temp.path().join("main.rs"), "fn main() {}\n").expect("write file");
+
+    let summary = scan_path(temp.path()).expect("scan");
+
+    assert!(
+        summary.health_score > 0,
+        "clean project should have positive health score"
+    );
 }
